# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CommissionKit API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Rep {
  id: string;
  name: string;
  email: string;
  role: string;
  planId: string | null;
  planName: string | null;
  createdAt: string;
}

export interface CreateRepBody {
  name: string;
  email: string;
  role: string;
  planId?: string | null;
}

export interface PlanTier {
  id: string;
  fromAmount: number;
  toAmount: number | null;
  /** Commission rate as a decimal (e.g. 0.05 = 5%) */
  rate: number;
}

export type PlanType = typeof PlanType[keyof typeof PlanType];


export const PlanType = {
  flat: 'flat',
  tiered: 'tiered',
  accelerator: 'accelerator',
} as const;

export interface Plan {
  id: string;
  name: string;
  type: PlanType;
  /** Used when type=flat */
  flatRate: number | null;
  /** Deal value at which accelerator kicks in */
  acceleratorThreshold: number | null;
  /** Rate applied above the accelerator threshold */
  acceleratorRate: number | null;
  /** Number of days before commission is clawed back if deal cancels */
  clawbackDays: number | null;
  tiers: PlanTier[];
  createdAt: string;
}

export interface CreatePlanTier {
  fromAmount: number;
  toAmount?: number | null;
  rate: number;
}

export type CreatePlanBodyType = typeof CreatePlanBodyType[keyof typeof CreatePlanBodyType];


export const CreatePlanBodyType = {
  flat: 'flat',
  tiered: 'tiered',
  accelerator: 'accelerator',
} as const;

export interface CreatePlanBody {
  name: string;
  type: CreatePlanBodyType;
  flatRate?: number | null;
  acceleratorThreshold?: number | null;
  acceleratorRate?: number | null;
  clawbackDays?: number | null;
  tiers?: CreatePlanTier[];
}

export type DealStage = typeof DealStage[keyof typeof DealStage];


export const DealStage = {
  closed_won: 'closed_won',
  closed_lost: 'closed_lost',
  pending: 'pending',
} as const;

export interface Deal {
  id: string;
  repId: string;
  repName: string;
  name: string;
  amount: number;
  closeDate: string;
  /** YYYY-MM */
  period: string;
  stage: DealStage;
  notes: string | null;
  createdAt: string;
}

export type CreateDealBodyStage = typeof CreateDealBodyStage[keyof typeof CreateDealBodyStage];


export const CreateDealBodyStage = {
  closed_won: 'closed_won',
  closed_lost: 'closed_lost',
  pending: 'pending',
} as const;

export interface CreateDealBody {
  repId: string;
  name: string;
  amount: number;
  closeDate: string;
  period: string;
  stage: CreateDealBodyStage;
  notes?: string | null;
}

export interface ImportDealsBody {
  /** YYYY-MM period for these deals */
  period: string;
  deals: CreateDealBody[];
}

export interface ImportDealsResult {
  imported: number;
  skipped: number;
  errors: string[];
}

export interface CommissionResult {
  id: string;
  repId: string;
  repName: string;
  dealId: string;
  dealName: string;
  dealAmount: number;
  rateApplied: number;
  commissionAmount: number;
  calculationNote: string;
}

export type CommissionRunStatus = typeof CommissionRunStatus[keyof typeof CommissionRunStatus];


export const CommissionRunStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface CommissionRun {
  id: string;
  period: string;
  totalCommission: number;
  totalDeals: number;
  repsCount: number;
  status: CommissionRunStatus;
  error?: string | null;
  createdAt: string;
}

export type CommissionRunWithResultsStatus = typeof CommissionRunWithResultsStatus[keyof typeof CommissionRunWithResultsStatus];


export const CommissionRunWithResultsStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface CommissionRunWithResults {
  id: string;
  period: string;
  totalCommission: number;
  totalDeals: number;
  repsCount: number;
  status: CommissionRunWithResultsStatus;
  error?: string | null;
  createdAt: string;
  results: CommissionResult[];
}

export interface CreateRunBody {
  /** YYYY-MM period to calculate */
  period: string;
}

export interface RepEarning {
  repId: string;
  repName: string;
  totalCommission: number;
  totalDeals: number;
  totalRevenue: number;
}

export interface DashboardSummary {
  period: string;
  totalCommission: number;
  totalRevenue: number;
  totalDeals: number;
  totalReps: number;
  repEarnings: RepEarning[];
  recentRuns: CommissionRun[];
}

export interface RepDealResult {
  dealId: string;
  dealName: string;
  dealAmount: number;
  closeDate: string;
  rateApplied: number;
  commissionAmount: number;
  calculationNote: string;
}

export type RepSummaryMonthlyHistoryItem = {
  period: string;
  totalCommission: number;
  totalDeals: number;
};

export interface RepSummary {
  repId: string;
  repName: string;
  email: string;
  planName: string | null;
  period: string;
  totalCommission: number;
  totalRevenue: number;
  totalDeals: number;
  dealBreakdown: RepDealResult[];
  monthlyHistory: RepSummaryMonthlyHistoryItem[];
}

export type ListDealsParams = {
repId?: string;
/**
 * Filter by period (YYYY-MM)
 */
period?: string;
};

export type GetRepSummaryParams = {
/**
 * Filter by period (YYYY-MM). Defaults to current period.
 */
period?: string;
};