# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CommissionKit API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reps:
    get:
      operationId: listReps
      tags: [reps]
      summary: List all sales reps
      responses:
        "200":
          description: List of reps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rep"
    post:
      operationId: createRep
      tags: [reps]
      summary: Create a sales rep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRepBody"
      responses:
        "201":
          description: Created rep
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rep"
  /reps/{id}:
    get:
      operationId: getRep
      tags: [reps]
      summary: Get a rep by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rep
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rep"
    put:
      operationId: updateRep
      tags: [reps]
      summary: Update a rep
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRepBody"
      responses:
        "200":
          description: Updated rep
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rep"
    delete:
      operationId: deleteRep
      tags: [reps]
      summary: Delete a rep
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Plans ─────────────────────────────────────────────────────────────────
  /plans:
    get:
      operationId: listPlans
      tags: [plans]
      summary: List all commission plans
      responses:
        "200":
          description: List of plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: createPlan
      tags: [plans]
      summary: Create a commission plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanBody"
      responses:
        "201":
          description: Created plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /plans/{id}:
    get:
      operationId: getPlan
      tags: [plans]
      summary: Get a plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    put:
      operationId: updatePlan
      tags: [plans]
      summary: Update a commission plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanBody"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    delete:
      operationId: deletePlan
      tags: [plans]
      summary: Delete a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Deals ─────────────────────────────────────────────────────────────────
...
  /dashboard/rep-summary/{repId}:
    get:
      operationId: getRepSummary
      tags: [dashboard]
      summary: Summary for a single rep including deal-by-deal breakdown
      parameters:
        - name: repId
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: query
          required: false
          schema:
            type: string
          description: "Filter by period (YYYY-MM). Defaults to current period."
      responses:
        "200":
          description: Rep summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepSummary"
components:
  schemas:
...
    Rep:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        planId:
          type: string
          nullable: true
        planName:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, email, role, planId, planName, createdAt]
    CreateRepBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        planId:
          type: string
          nullable: true
      required: [name, email, role]

    # ── Plan ─────────────────────────────────────────────────────────────────
    PlanTier:
      type: object
      properties:
        id:
          type: string
        fromAmount:
          type: number
        toAmount:
          type: number
          nullable: true
        rate:
          type: number
          description: "Commission rate as a decimal (e.g. 0.05 = 5%)"
      required: [id, fromAmount, toAmount, rate]
    Plan:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [flat, tiered, accelerator]
        flatRate:
          type: number
          nullable: true
          description: "Used when type=flat"
        acceleratorThreshold:
          type: number
          nullable: true
          description: "Deal value at which accelerator kicks in"
        acceleratorRate:
          type: number
          nullable: true
          description: "Rate applied above the accelerator threshold"
        clawbackDays:
          type: integer
          nullable: true
          description: "Number of days before commission is clawed back if deal cancels"
        tiers:
          type: array
          items:
            $ref: "#/components/schemas/PlanTier"
        createdAt:
          type: string
      required: [id, name, type, flatRate, acceleratorThreshold, acceleratorRate, clawbackDays, tiers, createdAt]
    CreatePlanTier:
      type: object
      properties:
        fromAmount:
          type: number
        toAmount:
          type: number
          nullable: true
        rate:
          type: number
      required: [fromAmount, rate]
    CreatePlanBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [flat, tiered, accelerator]
        flatRate:
          type: number
          nullable: true
        acceleratorThreshold:
          type: number
          nullable: true
        acceleratorRate:
          type: number
          nullable: true
        clawbackDays:
          type: integer
          nullable: true
        tiers:
          type: array
          items:
            $ref: "#/components/schemas/CreatePlanTier"
      required: [name, type]

    # ── Deal ─────────────────────────────────────────────────────────────────
...
    RepDealResult:
      type: object
      properties:
        dealId:
          type: string
        dealName:
          type: string
        dealAmount:
          type: number
        closeDate:
          type: string
        rateApplied:
          type: number
        commissionAmount:
          type: number
        calculationNote:
          type: string
      required: [dealId, dealName, dealAmount, closeDate, rateApplied, commissionAmount, calculationNote]
    RepSummary:
      type: object
      properties:
        repId:
          type: string
        repName:
          type: string
        email:
          type: string
        planName:
          type: string
          nullable: true
        period:
          type: string
        totalCommission:
          type: number
        totalRevenue:
          type: number
        totalDeals:
          type: integer
        dealBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/RepDealResult"
        monthlyHistory:
          type: array
          items:
            type: object
            properties:
              period:
                type: string
              totalCommission:
                type: number
              totalDeals:
                type: integer
            required: [period, totalCommission, totalDeals]
      required: [repId, repName, email, planName, period, totalCommission, totalRevenue, totalDeals, dealBreakdown, monthlyHistory]